feat(card): Implement cards load and image check
Function init_cards checks if cards file path exists
and whether it is a dir file, if it is, file path to
image and count is added for each card.

If we have one of each card, which we check with count
in previous function, we call check_card function,
which checks whether each card image has valid format
(fits terminal window). Card image is located in
assigned file path for that card.